tests/wpt/web-platform-tests/css/css-writing-modes/text-orientation-parsing-001.html

<!DOCTYPE html>
<title>CSS Writing Modes: parsing text-orientation</title>
<link rel="author" title="Koji Ishii" href="mailto:kojiishi@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-writing-modes-3/#text-orientation">
<meta name="assert" content="This test asserts the parser and getComputedStyle works correctly for the text-orientation property.">
<meta name="flags" content="dom">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div
  data-expected="mixed" title="text-orientation initial value"></div>

<div style="text-orientation: upright"
  data-expected="upright"></div>
<div style="text-orientation: sideways"
  data-expected="sideways"></div>

<div style="text-orientation: upright">
  <div
    data-expected="upright" title="text-orientation should inherit"></div>

  <div style="text-orientation: mixed"
    data-expected="mixed"></div>

  <div style="text-orientation: initial"
    data-expected="mixed"></div>
  <div style="text-orientation: inherit"
    data-expected="upright"></div>

  <div style="text-orientation: foo"
    data-expected="upright"></div>
</div>

<script>
Array.prototype.forEach.call(document.querySelectorAll("[data-expected]"), function (element) {
  test(function () {
    var actual = getComputedStyle(element).textOrientation;
    assert_equals(actual, element.dataset.expected);
  }, element.title || element.getAttribute("style"));
});
</script>
