LayoutTests/imported/w3c/web-platform-tests/css/css-writing-modes/text-orientation-parsing-001-expected.txt


FAIL text-orientation initial value assert_equals: expected (string) "mixed" but got (undefined) undefined
FAIL -webkit-text-orientation: upright assert_equals: expected (string) "upright" but got (undefined) undefined
FAIL -webkit-text-orientation: sideways assert_equals: expected (string) "sideways" but got (undefined) undefined
FAIL text-orientation should inherit assert_equals: expected (string) "upright" but got (undefined) undefined
FAIL -webkit-text-orientation: mixed assert_equals: expected (string) "mixed" but got (undefined) undefined
FAIL -webkit-text-orientation: initial assert_equals: expected (string) "mixed" but got (undefined) undefined
FAIL -webkit-text-orientation: inherit assert_equals: expected (string) "upright" but got (undefined) undefined
FAIL -webkit-text-orientation: foo assert_equals: expected (string) "upright" but got (undefined) undefined
